Remove keyisobj logic from indexor
The following didn't make any sense and wasn't used anywhere.

    keyisobj = type(key) != str

    for e in self._list:
        if keyisobj:
            if e is key:
                return e

I needed to index something with a UUID object as its .id proprety. Since this
was an object, this keyisobj logic was being invoked causing the value (not the
key) to be identity tested against the UUID object so it failed. I needed it to
be equality tested against the .id propery