[orm] Remove automatic deferred-loading
Previously, collection entities were loaded only when a presumed valid
attribute was called. This worked well until it was realized that one of the
determinations to load was whether or not the entities collection had been
contructed to load with a conditional (i.e, where object). See the following
line in the diff:

-        if not self.orm.where and not self.orm.ischunk:

Since it's obviously possbile to want to load a root entities collection that
doesn't have a condititonal, the ability to distinguish entities that needed to
be loaded from ones that didn't could not be made. Now, we use entities.load()
to explicitly load objects. Sadly, there doesn't seem to be a way to load entities
automatically given the right attribute is called.