[orm] Fix load/reload issue
Added a entities.reload() method so correct reloading could be done. An
assignment to set the isloaded to False in entities.clear() (see comments in
patch) was causing problems and was it was discovered that it shouldn't exist.
Removing the assignment caused several other small issues which were also fixed
by this patch.