[orm] Optimize orm.entity.__getattribute__
On reviewing the results of cProfile on test.py, it was discovered that
entity.__getattribute__ consumed most of the procesor resources  (aside
from database operations) to complete the script.

Extensive optimizations to entity.__getattribute__ were made in this
patch. The results where a 58% increase in the time it takes to run
test.py.

Before:
    real    1m19.278s
    user    1m10.284s
    sys     0m1.152s

After
    real    0m46.473s
    user    0m38.532s
    sys     0m1.072s