[gem] Allow pseudocollection to have multiple entity object
There was a bug discovered where appending a second entity object to a
pseudocollection would not work. This was because pseudocollection were
populated once when called and no logic was ever put in place to allow
for additional entity objects.

Also in this patch: work to make reflexive associations functional.