[gem] Use map.value to get pseudocollections
Using

    map.value.orm.entities.__name__

instead of

    map.entity.orm.entities.__name__

means we get the name of the actual entity type. This is important for
getting the pseudocollections of subentity objects.

For example: `map` might be the entitymapping `object` for the
`artist_artist` association. That map may have an `entity` attribute of
class `artist`. However, a `singer` object may be in the map.value
property in which case, we would want the `singers` pseudocollections,
not the `artists` pseudocollections.

Also added a bunch of unrelated comments.