[gem] Add guestbook to e._save()
I removed the follow* parameters from _save and replaced them with a
basic guestbook to prevent recursion.

I can't remember why I didn't use a guestbook in the first place. It
seems to have been an error in judgement.

Several of the tests failed after this patch, but it was
fairly easy to correct them all. The next commit will be the correction
to the tests needed by this patch.