[gem] Fix brokenrules issue
Logic was added to short circuit the collection of broken rules when an
associations collection has constituents that have already been visited.

Note that the first parameter to entities._getbrokenrules was called
`es`. However, this was being correctely used as the guestbook list so
the name was changed to guestbook. Logic to test the guestbook
was then added.