[gem] Test for collinear in pseudocollection creation
Adding the subtype painter to to a reflexive association along with the
subtype singer caused a problem when generating a pseudocollection. A
test for entity collinearity was correct this issue.

NOTE This ended up fixing some other issues. The following commit will
contain changes to the code that could be made now that this patch has
been committed.