[gem] Fix saving new graphs when .super is None
During the writing of a unit test, I failed to set some inherited orm
attributes on the entity I was testing. I should have gotten a
brokenrules exception to let me know that the attributes needed to be
set. However, setting these attributes was necessary for .super to be
set. With .super unset, the graph wasn't saved completely and no
exception was thrown. The solution was to make sure that the entity's
`_super` is always set when the subentity is new.