[web] Make menu.items inherit from dom.lis
Prior to this patch, menu.items inherited from both entities.entities
and ul. This made sense but the consequences of having an object inherit
from entities and entity (via ul) was too much for the framework to
handle (particularly entities.__init__ and entities.append). Because
this was causing so many problems, menu.items now inherits from dom.lis.
A few lines were added to menu.items.elements to ensure that it
expresses as a <ul> element to .html and .pretty.