[http] Made option.seletected a bool
Setting an attribute to None is enough to get it in the attribute list:

    <element attr>

However, selecting an attribute should be a boolean assign. So instead
of:

    opt.selected = None

we want to type

    opt.selected = True

This will probably be true of most attributes like this. It would be
nice to come up with a universal solution to this problem.