[http] Create default error page in dom.site
This begins the creation of error reporting for web pages.

This patch in particular deal with an HTTP exception being thrown that
has no corresponding page in the abstract site (dom.site) or the
concrete site (foonet) so if an exception like ImATeapotError is raised,
it will go to the most generic page '/en/error' for reporting to the
user.

Also in this patch:

    * Add HttpError.phrase
    * Make dom.site.pages a @property
    * Add site.__call__
    * Correct the way non-type-hinted parameters are determined
    * Add pom_page.it_raises_im_a_teapot