[ev] Automatically assign UUIDs to element.id
This patch makes it so that, by default, when an element is
instantiated, it's id attribute will have a base64 representation of a
UUID.

    p = dom.paragraph
    id = primative.uuid(base64 = p.id)
    assert type(id) is primative.uuid

Adding this feature broke a lot of unit tests which were expecting no id
attribute to exist by default.

A lot of other issue arose as well. For example, pom.menu and its
subsidiaries were not memoizing their subordinates which mean, for
example, to calls to menu.pretty would result in elements having
different id values.