[ev] Allow el[sels] to include el in it's matching
Previsously, the following CSS selector would not include the element
being selected on:

    ps = p['p']

Any dom.p's within `p` would be selected. However, `p` itself would not.
This patch attemps to fix this since it seems more reasonable that we
would want to subject `p` to the selection. A unit test is still failing
however, which is why the old `return` is still there.