[gem] Add fields to select
An issue keeps appearing where INSERTs fail because of ordinal parity
when the table is created in a different process. I haven't yet been
able to write a test that reproduces this, however.

This patch names fields in the INSERTs so row creation is independent of
the ordinality of fields in database.