[gem] Add **kwargs to primative.datetime.utcnow()
This is for getting an offset of now using familiar Python semantics:

    yesterday = primative.datetime.utcnow(days = -1)

There is also a now() override to prevent it from being used by
accident since non-utc dates will usually be a mistake. In the future,
it will return a datetime as long as a TZ argument has been provided.