[ev] Disallow appending to dom.text nodes
This should have been disallowed from the begining. All dom.text nodes
should be leafs.

This problem was noticed when trying to account for changes made to a
DOM. Since a "diff" can't be created to show that a text node has an
element attached to it (because text nodes can't have id attributes when
serialized into HTML strings), the ability had to be disallowed.

A few tests depended on the ability to append elements to text nodes.
They were corrceted in this patch.