[ev] Don't needlessly raise entities.onremove
It was noticed that entities.__setitem__ would raise the on remove event
to indicate that the previous item had been removed. However, the
previous item may have been the same item as the one being replace
therefore no removal (or addition) occurs. This problem can happen in
situations like:

    es.first = es.first

But it can also hapen in situations like

    els.first +=  dom.p()