[gem] Remove need to provide subject to association
For non-reflexive associations, a bug required you to make an assign to
the subjective side of an association. This is unnecessary because the
ORM can figure the subjective side out. Consider the diff below:

         paper.product_features += product.product_feature(
             type=product.product_feature.Optional,
             feature=product.quality(name='Extra glossy finish'),
-            product=paper  # TODO d3bd0e6d
         )

The comment for d3bd0e6d mentions the need to search down the inheritence
hierarchy but searching up the inheritence hierarchy was needed for the
examples I had in test.py so the isinstance() function was used. Perhaps
in the future, searching down the inheritence hierarchy will be
necessary.