[gem] Fix bug that caused constituents to not save
Note that the ORM bug seems to be fixed here. However, I was only able
to find two instance of the test.py code that was affected by this code.
There are other areas of test.py that were affected by this code (they
have TODO|FIXMEs), however that logic is currenently disable by a
premature `return` statement that was placed there on account of another
bug. Once that bug(s) becomes fixed, I can fix the other areas of
test.py that are affected by the ORM bug fixed in this commit.