[gem] Rewrote orm.entities.onadd
The original implementation of entities.onadd raised an error if the
number of composite's super entities didn't match the constiuents
entities. The way it was written - that it even cared about the
composite's superentities - didn't make sense to me so I rewrote it.

The rewrite fixes a bug that cause the error to be raised when
`gem.party.roles` exists. Now that the bug is fixed, the `roles`
attribute was uncommented and the code that used it was also uncommented
so to speak (actually, the premature `return`s was removed).