[gem] Allow e to access the composites of their supers
Code in entity.__getattribute__ prevented the return of composites that
were derived from an ascension up the inheritance tree. The reason given
was commented in the logic:

    if map_type is entitymapping:
        # We don't want an entitymapping from a super
        # returned.  This would mean conc.artist would
        # work. But concerts don't have artists;
        # presentations do. Concerts have singers.
        msg = "'%s' object has no attribute '%s'"
        msg %= self.__class__.__name__, attr
        raise AttributeError(msg)

This check prevented a lot of code from working as expected.  For
example, to get the the party composite defined at party.role, we had to
use orm.super twice to manually ascend the inheritance tree:

            <shipto>  <customer> <role>
    acme1 = shipto1.orm.super.orm.super.party

even though, we would natuarally expect the ascension to be done
automatically:

    acme1 = shipto1.party

The rational given in the comment in the above code snippet doesn't seem
to make sense on review. Why shouldn't a concert have an `artist`.
Shouldn't the following be true:

    assert conc.artist is conc.orm.super.artist

Removing the check solved the problem this was causing. Removing the
check only broke three tests and those only broke because they were
explicitly testing this logic.