[gem] Switch back to using pdb.set_trace
Aside from the syntax highlighting afforded by IPython's debugger, it
was not doing very much despite all of its neat features. It also made
the ``bt`` command print a stacktrace that was too verbose.

With the upgrade to Ubuntu 18, it was noticed that tab-completion did
not work with IPython's debugger. I switch the code back to the internal
pdb.set_trace and the problem was solved.