[mig] Return precision of 6 for datetime
The mapping.definition property returns 'datetime(6)` with a hardcoded
'6'. Hovever, the mapping.precision would return a None for datetimes.
This cause confusions with db.column:

    map = mapping(type=orm.types.datetime)
    assert map.precision is None # Should be 6
    assert map.definition == 'datetime(6)'  # This is probably right

    # Precision is None because map.precision is None
    assert db.column(map).definition == 'datatime(None)'

It's not clear why mapping didn't use the `precision` property. A little
more research into how we define datetime's and their precicions to the
database may be needed.