[file] Use getattr in orm._getargs
Prior to this patch, map.value was, for some reason, used to get the arg
values. This worked because each of the mappings are evaluated by the
default brokenrules before INSERTs and UPDATEs are executed.

Changing to getattr() doesn't change much, but in cases where the
default brokenrule logic isn't used, we can at least get the values in
the database correct.