Pass `cmp` to __setattr__
The option to compare (cmp) was not being passed up the inheritence
hierarchy thourgh orm.entity.__setattr__. However, sometimes we want the
supers to honor the request not to compare, so unnecessary SELECTs
aren't performed. Passing `cmp` to entity.__setattr__ makes this
possible.

This patch is, as of now, still considered experimental. Further tests
will confirm whether it will be kept, modified or undone.

On branch orm-specialize-elements