Add it_streams_specialized_entity_objects
Note that when testing, we don't sort. This is because adding the
subjoins caused an `ORDER BY id` to be ambigious due to the additional
fields being SELECTed from other tables. We need to devise a strategy to
call sort() and sorted() without being ambigous. That effort will likely
take place in this branch.

On branch orm-specialize-elements