Don't append proprietor filter for chunks.
Streamed entities will add a proprietor filter to themselves and then
pass the filter to the chunked entities. Prior to this commit, the
chunked entities would receive the proprietor filter and then add it's
own. This caused an innocuous bug where the WHERE predicates where being
issued to the database to filter the proprietor.

        WHERE (
            `m_pma_pr`.proprietor__partyid = _binary %s AND
            `m_pma_pr`.proprietor__partyid = _binary %s
        )

        args == [
            b'wW\xf0\x92\xe05Kk\xaf\x80\x89\x88\xd9\x94\x9dR',
            b'wW\xf0\x92\xe05Kk\xaf\x80\x89\x88\xd9\x94\x9dR'
        ]

This patch makes entities objects aware if they are chunked. Based on
that awareness, they will not append a proprietor filter, assuming one
has been passed in.