Add --rebuild argument to tester.cli
This allows a unit tester to add:

    if self.rebuildtables:
        ...

to any test code that DROPs and CREATEs table. However, so far, this
conditional logic has only been applied to test_orm.__init__.

Also, a reference to `testers` is passed to a tester's __init__. This
makes it possible for `self.rebuildtables` to be called from __init__.
Consequently, classes that inherit from `tester` now need to ensure that
all arguments to __init__ are past to the super().

    def __init__(self, *args, **kwargs):
        super().__init__(*args, **kwargs)

On branch orm-specialize-composite