Add some protection when propagating proprietor
Note that with the introduction of the new entity propogation logic
(in cc40dd684ac2ceb5ecef7d3f97c9c0480fc310b4), some other tests broke in
a good way. The proprietor started to be propagated up the tree as well,
which broke tests that were sort of flagging this problem. The changes
made to those tests are part of this commit as well.

On branch orm-specialize-composite