Restrict validation to only: type(self) is file
This is a temporary fix to reduce the number of times this database hit
was called. It was causing a simple save() to take about 20 seconds.

It raises an issue with the archetecture of brokenrules themselves. Part
of the problem was that when

    resx.isvaild

is called, resx.getbrokenrules ascends to file.getbrokenrules. This
causes file's validation to be done even though we are only interested
in resx's validity. The reason we are only interested in resx's validity
is because resx's super, file, will be tested later, resulting in
another call to the database.

There are other places that are testing resx's validity multiple times,
such as when entity.getbrokenrules() ascends. For this reason, and for
other reasons that still aren't clear, about 8000 SELECTs were being
issued here from a a call to resources.save() in a test, causing the
save() to take about 20 seconds.

Going forward, we need a design that does not require getbrokenrules to
call super().getbrokenrules(). We may need to revert to the `brokenrule`
@property and do inheritence ascension in the ORM.

On branch complete-file-api