Change entity._save to evaluate brs only if there is a mutation
See the XXX comments for more information in the patch.

Evaluating the broken rules even when a mutation wasn't needed was
causing problems because the inode.brokenrules checks if the inode
exists. Since an inode existed, a BrokenRulesError was raised when it
shouldn't have been, so this change was made.

The patch cleared up all of these problems, plus it made the tests in
testfile.file_file run 3 times as fast. So I like this patch. Hopefully,
it want cause problems when all the tests are run.

On branch cache-inodes