Swallow InterfaceError
After moving to MySQL 8, when tests in test.py wanted to kill the
connection (obviously, something that would probably not come up often
in a normal production environment), the MySQLdb driver would throw an
InterfaceError. Other times it would throw an OperationalError with an
error code of 2013. At this point, its difficult to understand what its
reasonings are.

On branch cache-inodes-3_8