Remove duplicate party.communication class
This had to be removed anyway because it was a duplicate. I removed it
as part of this branch in order to get all tests to succeed.

The presense of this class was causing a wierd error in orm.populate
when a `communication` object was being tested with isinstance():

    if isinstance(e, party.communication): ...

Appearently, this test didn't work correctly because the wrong
party.communication class was being referenced.

On branch cache-inodes