Remove pseudocollections from orm.py
This patch starts the process of removing pseudocollections from the ORM
(orm.py).

Pseudocollections were a way to make working with many-to-many
relationships easier by ignoring the association entity. For example,
instead of

    mov = movie
    act = actor

    # Create an association between the movie and the actor
    mov.movie_to_actors += movie_to_actor(actor=act)
    mov.save()

we could use the pseudocollection logic to write this:

    mov = movie
    act = actor

    # Create an association between the movie and the actor using less
    # code
    mov.actors += act
    mov.save()

After writting the general entity model, it was noticed that virtually
all association entities have significant attributes. For example in the
above, we would probably want an attribute in the association entity to
indicate what type of role the actor played in the movie, and when they
played that role

    mov.movie_to_actors += movie_to_actor(
        actor = act,
        role  = 'leading',
        start = startdate,
        end   = futuredate,
    )

The problem is, pseudocollection logic still needs to be maintained when
changes are made to the ORM's logic, and there may still be bugs in it.
Additionally, pseudocollections may be imposing a performance penalty
even when not being used. Since they no longer appear to provide much
value, this is an untenable maintenance burden, so we are removing them.
This patch can be used as a reference to reinstate them at a later date
if we change our mind. There will be more patches in this branch to get
all the tests running correctly; as stated: this patch only represents
the removal of all pseudocollection logic from orm.py

On branch remove-pseudoclasses