Change orm.attr.attr from function to class
It was noticed that when an imperative ends up calling another
imperative, the variables injected into the attr function would end up
being wrong. This was because attr() was being treated as a global
variable in a sense. The solution was to change attr to a callable class
so it could be instantiated and hold its own independent state values.

On branch remove-XXX-comments