Report MB used by testing process
This reports the number of MB used by the test process at the moment a
test is run.

    # 67M it_loads_alot_of_objects
    # 127M it_does_something_else
    # [...]

To get this value, we use

    import resource
    resource.getrusage(resource.RUSAGE_SELF).ru_maxrss

There was some reworking of ```testers`` as well. Specifically,
oninvoketest was changed to onbeforeinvoketest. Also, the event handler
for this event was defined as a private method on the ``cli`` class,
instead of in lambdas.

On branch memory-leak