Remove r argument from entities.append
The r argument maintained a entities.entities collection of elements
that were appended successfully to the collection (as opposed to those
that were discarded due to not being unique when the unique element was
True.

There were a lot of overrides to entities.append, so these had to be
changed as well.

It was noticed that the append operation was taking a while according to
the profiler, so it was decided that this seemingly expensive collection
be removed because it was never needed. If we ever need the feature it
was intended for, we would probably be better off using event handling.

Perodoxiclally, no diminution of execution time was observed. I think
the system is a little slow today (it's a VM so the actual machine is
probably shared).

570.72user 2.57system 10:51.04elapsed 88%CPU

On branch perf