Add benchmark.time(setup='pass')
This is passed directly to the setup parameter in timeit.Timer(). It
turns out, this was not needed. The setup callable only gets called once
for each invocation of the timer. I need to pass a callable that get
invoked before each repeatition of the `stmt`. I may add this feature to
timeit by subclassing it.

On branch perf