Add entity index toggling and turn off by default
This adds the ability to change whether an entities collection will
maintain indexes for fast lookups. The toggle is set to off so indexing
will, by default, not occur. Indexing was added for entities collection
that grew into the hundreds and thousands. This usually doesn't happen
in the OLTP systems that the framework is intended to handle. Indexing,
however, has the consequence of making appends to entities collections
slow. By removing indexing for most entities collections, we seem to
have made a significant performance improvement to the over test suite.

303.70user 2.11system 6:22.06elapsed 80%CPU (0avgtext+0avgdata
133232maxresident)k
15680inputs+5808outputs (94major+70880minor)pagefaults 0swaps

On branch perf