Remove entities.table from orm._getselect
It was observed that creating a table object to format SELECT statements
was taxing any operation that needed a SELECT to get data out of the
database. This was removed and replaced with an unformated string for a
SELECT clause. It seems to have shaved another minute off total
runtime of the test script.

69user 1.96system 5:35.91elapsed 77%CPU (0avgtext+0avgdata
127016maxresident)k
14352inputs+5808outputs (62major+63194minor)pagefaults 0swaps
j

On branch perf