Remove automatic assignment of UUID to element id
Assigning id's to every element in a DOM was found to cause excessive
work because of a number of different operations:

    * The attributes collection had to be instantiated
    * The id attribute had to be instantiated
    * THe id attribute had to be appended

I don't think we will want this feature anyway because of the way
page-patching technology has been reenvisioned to work.

It was also discovered that when an id attribute had not been
set, it would return None. This was considered correct behavior, but
when some elements were cloned, these None values were used to set the
ids of new elements. Setting an attribute to None means the attribute
is set like a boolean attribute, which cause a bug.

    assert anchor.id is None
    assert '<a></a>' == anchor.html

    new_anchor.id = anchor.id

    #           ,------- boolean attribute
    assert '<a id></a>' == new_anchor.html

You can see that this had to be worked around in the clone methods. We
should come up with better semantics for these attribute values.

On branch perf