Make carapacian_com.site.Proprietor a @property
When test.py imports carapacian_com, the line

    Proprietor = party.company.carapacian

was executed. This caused the ORM to cache entity classes which caused
problems, particularly with the tests on abbreviations. ORM code
shouldn't be executed until all GEM classes have been discovered by the
interpreter. Wrapping this code in a @property prevents this.

On branch reify-add-orm_classproperty