Remove *args from dom.element.__init__
This was used as arguments to parameterize the body of an element. I'm
not sure why I thought this would be a good idea when I added it several
years ago. When I sent a bunch of CSS to dom.style, the appearently was
some text that looked like substitutable symbols, so the construtor
raise an exception. If we want string interpolation in the `body`
parameter we should just use an f-string.

On branch javascript-spa