Guess mime type of self.name
self.path was unnecessary input. However, it was also exposing another
bug that is fix in a seperate patch. This was partly due to the fact
that self.path does a lot of work to assemble its return value. Using
self.name just takes a lot less work.

On branch favicon