Use 'nav ul>li a'
We wanted 'nav>ul>li a' to work but there appears to be a bug in CSS
selection at the moment.

The change arise from the need to add an 'click' event handler to
submenu links; not just the links that are found directly off nav>ul>li.

On branch load-subpage-with-spa