Get the <main> tag for the replacement
Prior to this patch, we were replacing the existing main tag with an
`elements` collection that contained the <main> element. This worked,
but the <main> element would end up not having its `parent` attribute
assigned, which caused problem when a unit tests tried to re-GET the
same page.

On branch load-subpage-with-spa