Remove the notion of language code from pom
pom.py should not have to deal with language codes in URLs

    example.com/en/sub/page

www.py should deal with this data exclusively. The pom.pages collection
shouldn't care. This patch works considerably (if not completely) toward
that.

However, the site.languages property still needs to exist for its own
reasons. It was change to a @classproperty since it doesn't use `self`.
`foonet` needed to override this @classproperty because it supports
'es'.

Additionally, some centralized striping logic was added and used to
remove the language code from the path.

On branch load-subpage-with-spa