Make menu.items a ul instead of a lis
This was done to allow `menu.items.elements` to not require cloning.
Cloning here causes lots of problems. For example, when using a CSS
selector to get the anchor elements from the menu's indexor, editing the
achor object did not change the `menu`'s DOM because the anchors were
clones. Inheriting from `ul` and using the collection interface from
`element` allows us to not have to override `elements`. Thus, a lot of
code was able to be removed as well.

On branch data-spa-path