Fix bug parsing HTML with entities
HTML with HTML entities were not being processed correctly.

The solution I implemented removed the ability to preserve the HTML
entities when they are being parsed. This means, for example, if the
HTML string being parsed has &copy; in it, the resulting DOM would have
a © in it instead.  This may be good for now but we may need to fix this
later. There is a TODO in the patch that explains the problem a little
more.

On branch new-ticket